+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Limit data fetched by using map bounds
+
+### Fixed
 - Fixed search bar visuals
 
 ## [0.13.1] - 2025-03-10